requirements.txt
arxiv==2.1.3
pandas==2.2.2
pyarrow
PyYAML==6.0.1
numpy==1.26.4